+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 1.3.6 (unreleased)
 ------------------
 
+- #2108 Replace ParentAnalysisRequest ReferenceField by UIDReferenceField (#2069 port)
 - #2107 Replace Worksheet's Analysis ReferenceField by UIDReferenceField (#2022 port)
 - #2106 Skip Auditlog catalog if disabled for DX types multiplexer (#2094 port)
 - #2106 Allow to disable global Auditlogging (#2074 port)